Eighth-audit P0s: decouple LittleFS from the safety loop, atomic swap, strum id
Addresses the six P0 findings from the eighth audit.

1. LittleFS off the safety loop (P0-1): the shared g_stateMutex (which loop()
   takes for SysEx / status / reloads) no longer covers flash writes. A distinct
   g_storageMutex now guards all LittleFS access (save/load/delete/read/list);
   loop() NEVER takes it, so a long web-triggered flash write can't stall the
   E-stop / LIMIT / motion checks. g_stateMutex is only ever held for short
   in-memory work.

2. Atomic profile swap (P0-2): g_profile = p and applyProfile() now run under one
   StateGuard, so a web read never sees the profile assigned but the runtime not
   yet rebuilt.

3. Shared-strum chord identity (P0-3): notes started together get a strum-group
   id (one per automatic chord flush; explicit/triggered notes get their own).
   The barrier waits only for members of the SAME group, so a prepared-but-not-
   triggered or unrelated note never blocks the sweep and the fast-string /
   held-note cross-talk is gone.

4. Move watchdog (P0-4): a carriage that does not reach its fret within
   kMoveTimeoutMs (refused/stalled step command, stop far from target) faults the
   axis instead of waiting forever in MovingToFret.

5. Signed effective open note (P0-5): StringSpec::openNote is now a signed,
   unclamped int16 built from open note + capo + transpose. A reference below 0 /
   above 127 keeps exact fret math instead of clamping to a wrong open string.

6. Old fingers up before profile teardown (P0-6): profile activation is now two
   phase — phase 1 stops the motors and drives the OLD servos to rest (keeping
   servo power), phase 2 (after they lift) tears down, swaps, and re-homes. A
   profile that removes/reassigns a finger servo can no longer leave a finger
   pressed while the new profile homes. A panic / E-stop cancels a pending
   activation.

Also serialises the GET /api/profiles slot list under the storage lock.

Tests: +negative-effective-open-note. 120 tests / 2211 checks, 0 failures; all
ESP32 TUs pass the host compile-check.

+++ firmware/src/core/instrument/NoteAllocator.h
@@ -20,7 +20,10 @@ enum class SaturationStrategy : uint8_t {
 };
 
 struct StringSpec {
-    uint8_t openNote = 0;   // MIDI note of the open string
+    // Effective open pitch (open note + capo + transpose). SIGNED and NOT clamped
+    // to the MIDI domain: a transposed reference can legitimately fall below 0 or
+    // above 127, and clamping it would corrupt the fret math (audit P0-5).
+    int16_t openNote = 0;
     uint8_t maxFret = 12;
     bool enabled = true;
     bool faulted = false;
